Enum sui_source_validation::ValidationMode

pub enum ValidationMode {
    Deps,
    Root {
        deps: bool,
        at: Option<AccountAddress>,
    },
}

Details of what to verify

Variants

Deps

Validate only the dependencies

Root

Validate the root package, and its linkage.

Fields

deps: bool

Additionally validate the dependencies, and make sure the runtime and storage IDs in dependency source code matches the root package’s on-chain linkage table.

at: Option<AccountAddress>

Look for the root package on-chain at the specified address, rather than the address in its manifest.

Implementations

impl ValidationMode

pub fn deps() -> Self

Only verify that source dependencies match their on-chain versions.

pub fn root() -> Self

Only verify that the root package matches its on-chain version (requires that the root package is published with its address available in the manifest).

pub fn root_at(address: AccountAddress) -> Self

Only verify that the root package matches its on-chain version, but override the location to look for the root package to address.

pub fn root_and_deps() -> Self

Verify both the root package and its dependencies (requires that the root package is published with its address available in the manifest).

pub fn root_and_deps_at(address: AccountAddress) -> Self

Verify both the root package and its dependencies, but override the location to look for the root package to address.